requirements.txt
APScheduler==3.8.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-APScheduler==1.12.3
flask-mongoengine==1.0.0
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
pycodestyle==2.8.0
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
